# openapi.yaml
openapi: 3.0.3
info:
  title: FleetForge Scheduler API
  description: >
    REST control-plane API for the FleetForge distributed build scheduler.
    Worker registration and heartbeats are primarily transported over the
    gRPC bidirectional stream described in doc 1 (lower overhead at 10k-worker
    scale); the REST equivalents below exist for (a) workers/environments that
    cannot use gRPC, (b) tooling, dashboards, and CLIs, and (c) parity/testing.
    Both transports write to the same underlying state machine and are
    mutually consistent.
  version: "1.0.0"
  contact:
    name: FleetForge Platform Team
servers:
  - url: https://scheduler.internal.launchverse.dev/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local (docker-compose)
...
security:
  - bearerAuth: []
  - mtls: []
paths:
...
  /workers/{workerId}/drain:
    post:
      tags: [workers]
      summary: Begin draining a worker
      description: >
        Transitions worker to DRAINING. No new jobs will be assigned. Worker
        finishes its current job (if any) and is then automatically
        transitioned to OFFLINE by the scheduler. Returns immediately;
        draining is asynchronous.
      operationId: drainWorker
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        '202':
          description: Drain initiated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
        '404':
          description: Worker not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Worker already OFFLINE or DEAD
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workers/{workerId}/resume:
    post:
      tags: [workers]
      summary: Resume a drained or offline worker back to READY
      operationId: resumeWorker
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        '200':
          description: Worker resumed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
        '404':
          description: Worker not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Worker is DEAD; cannot resume, must re-register
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workers:
    get:
      tags: [workers]
      summary: List workers
      operationId: listWorkers
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkerStatus'
        - name: label
          in: query
          description: "Filter by label, format key=value. Repeatable."
          schema: { type: string }
          example: "arch=arm64"
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
      responses:
        '200':
          description: Paginated worker list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerList'
...
  /jobs:
    post:
      tags: [jobs]
      summary: Submit a build job
      operationId: submitJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmissionRequest'
      responses:
        '202':
          description: Job accepted and queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '400':
          description: Invalid job spec
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: "Duplicate idempotency_key: returns the original job instead of creating a new one"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
    get:
      tags: [jobs]
      summary: List jobs
      operationId: listJobs
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/JobStatus' }
        - name: repository
          in: query
          schema: { type: string }
        - name: worker_id
          in: query
          schema: { type: string, format: uuid }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
      responses:
        '200':
          description: Paginated job list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobList' }
  /jobs/{jobId}:
    get:
      tags: [jobs]
      summary: Get a single job
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job detail (includes retry history)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404':
          description: Job not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [jobs]
      summary: Cancel a job
      description: >
        If QUEUED, removed from the queue immediately. If ASSIGNED/RUNNING,
        a cancellation signal is sent to the worker over the gRPC stream;
        the job transitions to CANCELLED once the worker acknowledges (or
        after a force-cancel timeout, at which point the worker is presumed
        to have killed the process and the job is marked CANCELLED anyway).
      operationId: cancelJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Cancellation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404':
          description: Job not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Job already in a terminal state (SUCCESS/FAILED/CANCELLED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT for human/CI clients (job submission, admin actions)
    mtls:
      type: mutualTLS
      description: Client certificate auth for worker agents (registration, heartbeat)
  parameters:
    WorkerId:
      name: workerId
      in: path
      required: true
      schema: { type: string, format: uuid }
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    WorkerStatus:
      type: string
      enum: [REGISTERING, READY, BUSY, DRAINING, OFFLINE, DEAD]
    JobStatus:
      type: string
      enum: [QUEUED, ASSIGNED, RUNNING, SUCCESS, FAILED, RETRYING, CANCELLED]
...
    Worker:
      type: object
      properties:
        id: { type: string, format: uuid }
        hostname: { type: string }
        os: { type: string }
        cpu_cores: { type: integer }
        memory_mb: { type: integer }
        labels: { type: object, additionalProperties: { type: string } }
        capabilities: { type: array, items: { type: string } }
        status: { $ref: '#/components/schemas/WorkerStatus' }
        current_job_id: { type: string, format: uuid, nullable: true }
        last_heartbeat: { type: string, format: date-time }
        version: { type: string }
        available_capacity: { type: integer }
        epoch: { type: integer, format: int64 }
        registered_at: { type: string, format: date-time }
    WorkerList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Worker' } }
        next_cursor: { type: string, nullable: true }
        total_count: { type: integer }
    JobSubmissionRequest:
      type: object
      required: [repository, branch, commit_sha]
      properties:
        priority:
          type: integer
          minimum: 0
          maximum: 9
          default: 5
          description: "0 = highest priority, 9 = lowest."
        repository: { type: string, example: "launchverse/api" }
        branch: { type: string, example: "main" }
        commit_sha: { type: string, example: "a1b2c3d4e5f6" }
        labels: { type: object, additionalProperties: { type: string } }
        required_capabilities:
          type: array
          items: { type: string }
          example: ["docker", "gpu"]
        max_retries: { type: integer, default: 3 }
        idempotency_key:
          type: string
          description: Client-supplied key to safely retry job submission without creating duplicates.
    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        priority: { type: integer }
        repository: { type: string }
        branch: { type: string }
        commit_sha: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        required_capabilities: { type: array, items: { type: string } }
        retries: { type: integer }
        max_retries: { type: integer }
        status: { $ref: '#/components/schemas/JobStatus' }
        worker_id: { type: string, format: uuid, nullable: true }
        log_ref:
          type: string
          nullable: true
          description: Pointer to externally-stored logs (object storage URI).
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        retry_history:
          type: array
          items: { $ref: '#/components/schemas/RetryAttempt' }
    RetryAttempt:
      type: object
      properties:
        attempt_number: { type: integer }
        status: { $ref: '#/components/schemas/JobStatus' }
        worker_id: { type: string, format: uuid, nullable: true }
        error_message: { type: string, nullable: true }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
    JobList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Job' } }
        next_cursor: { type: string, nullable: true }
        total_count: { type: integer }
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        request_id: { type: string }